docs(search-input): optionsを別ファイルに切り出す
